Jieko

Documentation for Jieko.

Jieko.INTERFACE

Jieko.INTERFACE_LIST

Jieko.InterfaceMethod

Jieko.TypeAnnotation

Jieko.WhereParam

Jieko.interfaces

Jieko.@export_all_interfaces

Jieko.@interface

Jieko.INTERFACE — Constant

const INTERFACE = InterfaceSignature()

Similar to SIGNATURES but has more precise method information obtained directly from the @interface macro.

Jieko.INTERFACE_LIST — Constant

const INTERFACE_LIST = InterfaceList()

List all the interface methods of a module. It shows nothing if the binded object is not a module.

Jieko.InterfaceMethod — Type

struct InterfaceMethod

The signature information of an interface method.

Note

This is only used for printing for now. Thus we do not store the actual type signature at the moment. We should think about how to store the type signature and use this info in the future.

Jieko.TypeAnnotation — Type

struct TypeAnnotation

The annotation information of a type. This also stores the name user typed in the annotation, e.g an alias of type is used instead of the original type name. This is useful for printing the type annotation in docstrings.

Jieko.WhereParam — Type

struct WhereParam

The type variable information of a function.

Jieko.interfaces — Method

public Jieko.interfaces(mod::Module)

Return the interface stub storage for a module.

Jieko.@export_all_interfaces — Macro

@export_all_interfaces

Wrap all interfaces from the module in which this macro is called in a Prelude module and export them. This is useful for exporting all interfaces from a module to be used in other modules.

The concept of Prelude is borrowed from rust, where it is used to give users an explicit way to import all the symbols from a package. This is useful for the package authors to provide a default set of symbols to be used by the users via using MyPackage.Prelude without giving them the easy way of importing everything from the package via using MyPackage.

Because Prelude module only contains the API symbols, it makes it easier for the toolchain to check the APIs without mixing them with the implementation details.

Jieko.@interface — Macro

@interface <function definition>

Mark a method definition as interface. This will help the toolchain generating docstring and other things. The interface method is usually the most generic one that errors.